[Swift in WebKit] Work towards modularizing JSC private headers (Part 2) https://bugs.webkit.org/show_bug.cgi?id=298495 rdar://159990396
Reviewed by NOBODY (OOPS!).

The private interface of JSC should not expose the interface or implementation of FastJITPermissions.h.
Moreover, doing so fails module verification due to some malformed system headers the file depends on.

To fix, make FastJITPermissions.h project-level instead of private-level. Because JSC has a lot of inline
functions, this requires some changes in the dependencies of the header; specifically:

1. For each private header in JSC that directly or transitively depends on FastJITPermissions.h,
    (a) Create an associated "*Internal.h" header file at project-level
    (b) Identify all inline functions in the dependent header that depend on a function in FastJITPermissions.h
        (i) For each of these functions, move the function definition to the "*.Internal.h" file, and explicitly mark it as inline.
2. For each private header that depends on any of the functions that were moved to "*.Internal.h",
    (a) repeat the steps in (1)
3. For each implementation (cpp) file, and each project-level header file, that depends on any of those functions, include the "*.Internal.h" file.

The end result of this operation is the addition of several new *.Internal.h files, which better separates
the project interface of JSC with the private interface. All previous inlined functions remain inlined.

* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/assembler/ARM64Assembler.h:
* Source/JavaScriptCore/assembler/ARM64AssemblerInternal.h: Added.
(JSC::ARM64Assembler::linkPointer):
(JSC::ARM64Assembler::replaceWithVMHalt):
(JSC::ARM64Assembler::replaceWithJump):
(JSC::ARM64Assembler::replaceWithNops):
(JSC::ARM64Assembler::fillNops):
(JSC::ARM64Assembler::fillNearTailCall):
(JSC::ARM64Assembler::repatchPointer):
(JSC::ARM64Assembler::setPointer):
(JSC::ARM64Assembler::linkCompareAndBranch):
(JSC::ARM64Assembler::linkConditionalBranch):
(JSC::ARM64Assembler::linkTestAndBranch):
(JSC::ARM64Assembler::linkJumpOrCall):
* Source/JavaScriptCore/assembler/AbstractMacroAssembler.h:
(JSC::AbstractMacroAssembler::linkPointer): Deleted.
(JSC::AbstractMacroAssembler::repatchPointer): Deleted.
* Source/JavaScriptCore/assembler/AbstractMacroAssemblerInternal.h: Copied from Source/JavaScriptCore/dfg/DFGJumpReplacement.cpp.
(JSC::AbstractMacroAssembler<AssemblerType>::repatchPointer):
(JSC::AbstractMacroAssembler<AssemblerType>::linkPointer):
* Source/JavaScriptCore/assembler/AssemblerCommon.h:
(JSC::machineCodeCopy): Deleted.
* Source/JavaScriptCore/assembler/AssemblerCommonInternal.h: Copied from Source/JavaScriptCore/dfg/DFGJumpReplacement.cpp.
(JSC::machineCodeCopy):
* Source/JavaScriptCore/assembler/LinkBuffer.cpp:
* Source/JavaScriptCore/assembler/LinkBuffer.h:
(JSC::LinkBuffer::link): Deleted.
(JSC::LinkBuffer::patch): Deleted.
* Source/JavaScriptCore/assembler/LinkBufferInternal.h: Added.
(JSC::LinkBuffer::link):
(JSC::LinkBuffer::patch):
* Source/JavaScriptCore/assembler/MacroAssemblerARM64.h:
(JSC::MacroAssemblerARM64::replaceWithVMHalt): Deleted.
(JSC::MacroAssemblerARM64::replaceWithJump): Deleted.
(JSC::MacroAssemblerARM64::replaceWithNops): Deleted.
(JSC::MacroAssemblerARM64::repatchCall): Deleted.
(JSC::MacroAssemblerARM64::linkCall): Deleted.
* Source/JavaScriptCore/assembler/MacroAssemblerARM64Internal.h: Added.
(JSC::MacroAssemblerARM64::linkCall):
(JSC::MacroAssemblerARM64::repatchCall):
(JSC::MacroAssemblerARM64::replaceWithVMHalt):
(JSC::MacroAssemblerARM64::replaceWithJump):
(JSC::MacroAssemblerARM64::replaceWithNops):
* Source/JavaScriptCore/bytecode/CallLinkInfo.cpp:
* Source/JavaScriptCore/bytecode/Repatch.cpp:
* Source/JavaScriptCore/bytecode/StructureStubInfo.cpp:
* Source/JavaScriptCore/dfg/DFGJITCompiler.cpp:
* Source/JavaScriptCore/dfg/DFGJumpReplacement.cpp:
* Source/JavaScriptCore/dfg/DFGLazyJSValue.cpp:
* Source/JavaScriptCore/dfg/DFGSpeculativeJIT64.cpp:
* Source/JavaScriptCore/ftl/FTLExceptionTarget.cpp:
* Source/JavaScriptCore/ftl/FTLLowerDFGToB3.cpp:
* Source/JavaScriptCore/ftl/FTLSlowPathCall.cpp:
* Source/JavaScriptCore/jit/ExecutableAllocator.cpp:
* Source/JavaScriptCore/jit/ExecutableAllocator.h:
(JSC::performJITMemcpy): Deleted.
* Source/JavaScriptCore/jit/ExecutableAllocatorInternal.h: Added.
(JSC::performJITMemcpy):
(JSC::isJITPC):
* Source/JavaScriptCore/jit/JIT.cpp:
* Source/JavaScriptCore/jit/JITOpcodes.cpp:
* Source/JavaScriptCore/jit/ThunkGenerators.cpp:
* Source/JavaScriptCore/wasm/WasmBBQJIT.cpp:
* Source/JavaScriptCore/wasm/WasmThunks.cpp:
* Source/JavaScriptCore/yarr/YarrJIT.cpp:
