+++ Source/JavaScriptCore/assembler/AssemblerCommon.h
@@ -423,25 +423,7 @@ static ALWAYS_INLINE void* memcpyAtomicIfPossible(void* dst, const void* src, si
 }
 
 template<RepatchingInfo repatch>
-void* performJITMemcpy(void* dst, const void* src, size_t n);
-
-template<RepatchingInfo repatch>
-ALWAYS_INLINE void* machineCodeCopy(void* dst, const void* src, size_t n)
-{
-    static_assert(!(*repatch).contains(RepatchingFlag::Flush));
-    if constexpr (is32Bit()) {
-        // Avoid unaligned accesses.
-        if (WTF::isAligned(dst, n))
-            return memcpyAtomicIfPossible(dst, src, n);
-        return memcpyTearing(dst, src, n);
-    }
-    if constexpr ((*repatch).contains(RepatchingFlag::Memcpy) && (*repatch).contains(RepatchingFlag::Atomic))
-        return memcpyAtomic(dst, src, n);
-    else if constexpr ((*repatch).contains(RepatchingFlag::Memcpy))
-        return memcpyAtomicIfPossible(dst, src, n);
-    else
-        return performJITMemcpy<repatch>(dst, src, n);
-}
+ALWAYS_INLINE void* machineCodeCopy(void* dst, const void* src, size_t n);
 
 WTF_ALLOW_UNSAFE_BUFFER_USAGE_END
