+++ Source/JavaScriptCore/ftl/FTLExceptionTarget.cpp
@@ -28,7 +28,7 @@
 
 #if ENABLE(FTL_JIT)
 
-#include "LinkBuffer.h"
+#include "LinkBufferInternal.h"
 
 namespace JSC { namespace FTL {
