+++ Source/JavaScriptCore/wasm/WasmThunks.cpp
@@ -33,7 +33,7 @@
 #include "JSCJSValue.h"
 #include "JSInterfaceJIT.h"
 #include "JSWebAssemblyInstance.h"
-#include "LinkBuffer.h"
+#include "LinkBufferInternal.h"
 #include "ProbeContext.h"
 #include "ScratchRegisterAllocator.h"
 #include "WasmExceptionType.h"
